+++ src/aux-runtime/runtime/AuxLibrary.ts
@@ -470,8 +470,6 @@ import {
 import { Fragment, h } from 'preact';
 import htm from 'htm';
 import { fromByteArray, toByteArray } from 'base64-js';
-import type { Tester } from '@casual-simulation/expect';
-import expect, { iterableEquality } from '@casual-simulation/expect';
 import {
     parseRecordKey,
     isRecordKey as calcIsRecordKey,
@@ -2839,86 +2837,6 @@ export interface GetRoomRemoteOptionsFailure {
     remoteId: string;
 }
 
-const botsEquality: Tester = function (first: unknown, second: unknown) {
-    if (isRuntimeBot(first) && isRuntimeBot(second)) {
-        expect(getBotSnapshot(first)).toEqual(getBotSnapshot(second));
-        return true;
-    }
-    return undefined;
-};
-
-expect.extend({
-    toEqual(received: unknown, expected: unknown) {
-        // Copied from https://github.com/facebook/jest/blob/7bb400c373a6f90ba956dd25fe24ee4d4788f41e/packages/expect/src/matchers.ts#L580
-        // Added the testBots matcher to make testing against bots easier.
-        const matcherName = 'toEqual';
-        const options = {
-            comment: 'deep equality',
-            isNot: this.isNot,
-            promise: this.promise,
-        };
-
-        const pass = this.equals(received, expected, [
-            botsEquality,
-            iterableEquality,
-        ]);
-
-        const message = pass
-            ? () =>
-                  this.utils.matcherHint(
-                      matcherName,
-                      undefined,
-                      undefined,
-                      options
-                  ) +
-                  '\n\n' +
-                  `Expected: not ${this.utils.printExpected(expected)}\n` +
-                  (this.utils.stringify(expected) !==
-                  this.utils.stringify(received)
-                      ? `Received:     ${this.utils.printReceived(received)}`
-                      : '')
-            : () =>
-                  this.utils.matcherHint(
-                      matcherName,
-                      undefined,
-                      undefined,
-                      options
-                  ) +
-                  '\n\n' +
-                  this.utils.printDiffOrStringify(
-                      expected,
-                      received,
-                      'Expected',
-                      'Received',
-                      this.expand !== false
-                  );
-
-        // Passing the actual and expected objects so that a custom reporter
-        // could access them, for example in order to display a custom visual diff,
-        // or create a different error message
-        return { actual: received, expected, message, name: matcherName, pass };
-    },
-});
-
-function getBotSnapshot(bot: Bot) {
-    let b = {
-        id: bot.id,
-        space: bot.space,
-        tags:
-            typeof bot.tags.toJSON === 'function'
-                ? bot.tags.toJSON()
-                : bot.tags,
-    } as Bot;
-
-    let masks = isRuntimeBot(bot)
-        ? bot[GET_TAG_MASKS_SYMBOL]()
-        : cloneDeep(bot.masks ?? {});
-    if (Object.keys(masks).length > 0) {
-        b.masks = masks;
-    }
-    return b;
-}
-
 /**
  * Defines an interface that represents the set of additional options that can be provided when recording a file.
  */
@@ -3518,8 +3436,6 @@ export function createDefaultLibrary(context: AuxGlobalContext) {
             clearWatchBot,
             clearWatchPortal,
             assert,
-            assertEqual,
-            expect,
 
             html,
 
@@ -4433,40 +4349,6 @@ export function createDefaultLibrary(context: AuxGlobalContext) {
                 throw new Error('Assertion failed.');
             }
         }
-    }
-
-    function getAssertionValue(value: any) {
-        if (value instanceof Error) {
-            return value.toString();
-        }
-        return value;
-    }
-
-    /**
-     * Verifies that the given values are equal to each other.
-     * If they are not, then an error is thrown.
-     * This function is useful for automated testing since tests should ideally throw an error if the test fails.
-     * It can also be useful to make sure that some important code is only run if a precondition is met.
-     *
-     * @param first The first value to test.
-     * @param second The second value to test.
-     *
-     * @example Assert that the tag color is "blue"
-     * assertEqual(tags.color, "blue");
-     *
-     * @example Assert that the bot contains some specific tag values
-     * assertEqual(tags, {
-     *     color: "blue",
-     *     home: true,
-     *     homeX: 0,
-     *     homeY: 0
-     * });
-     *
-     * @dochash actions/debuggers
-     * @docname assertEqual
-     */
-    function assertEqual(first: any, second: any) {
-        expect(first).toEqual(second);
     }
 
     /**
@@ -17043,6 +16925,8 @@ export function createDefaultLibrary(context: AuxGlobalContext) {
      * If null is provided, then a seed will be chosen in a somewhat unpredictable manner.
      *
      * @example Set the random seed for math.random() and math.randomInt().
+     * import expect from 'expect';
+     *
      * math.setRandomSeed(123);
      *
      * expect(math.randomInt(0, 10)).toBe(9);